(BSR)[PRO] chore: remove old quantityInput and TextInput
