+++ pro/src/components/FormLayout/FormLayout.stories.tsx
@@ -2,7 +2,7 @@ import type { StoryObj } from '@storybook/react'
 
 import { Button } from 'ui-kit/Button/Button'
 import { ButtonVariant } from 'ui-kit/Button/types'
-import { TextInput } from 'ui-kit/form/TextInput/TextInput'
+import { TextInput } from 'ui-kit/formV2/TextInput/TextInput'
 
 import { FormLayout, FormLayoutProps } from './FormLayout'
