--- pro/src/ui-kit/form/QuantityInput/QuantityInput.mdx
@@ -1,30 +1,0 @@
-{/* QuantityInput.mdx */}
-
-import { Meta, Primary, Controls, Story } from '@storybook/blocks'
-
-import * as QuantityInputStories from './QuantityInput.stories'
-
-<Meta of={QuantityInputStories} />
-
-# QuantityInput
-
-A QuantityInput component is a combinaison of a TextInput and a Checkbox to define **quantities**. This component should be used when undefined quantity is meant to be interpreted as unlimited.
-
-## Using QuantityInput over TextInput
-
-```jsx
-<TextInput
-  label="Quantity"
-  type="number"
-  value={quantity}
-  onChange={setQuantity}
-/>
-```
-
-If undefined quantity is meant to be interpreted as unlimited, the QuantityInput component should always be used.  
-Otherwise, using TextInput is fine enough (mind min, max, and step props attributes).
-
-## Extending props
-
-The QuantityInput component picks a few props from the TextInput component, which itself sees its props extended from FieldLayoutBase and BaseInput.  
-It's best to choose what needs to be picked than extending the props from the TextInput component so the QuantityInput component can be more flexible and easier to maintain.
