--- pro/src/ui-kit/form/TextInput/TextInput.mdx
@@ -1,18 +1,0 @@
-{/* TextInput.mdx */}
-
-import { Meta, Primary, Controls, Story } from '@storybook/blocks';
- 
-import * as TextInputStories from './TextInput.stories';
- 
-<Meta of={TextInputStories} />
- 
-# TextInput
-
-A TextInput component is wrapped with a FieldLayout component and built on top of the BaseInput component.  
-As such, it inherits all the props from BaseInput and FieldLayout.
-
-<Story of={TextInputStories.Default} />
-
-## Props
-
-<Controls />
