--- pro/src/ui-kit/form/TextInput/TextInput.module.scss
@@ -1,20 +1,0 @@
-// @use 'styles/mixins/_forms.scss' as forms;
-@use "styles/mixins/_rem.scss" as rem;
-
-.text-input {
-  &-readonly {
-    display: flex;
-    align-items: center;
-    min-height: rem.torem(40px);
-  }
-
-  &-group {
-    display: flex;
-    align-items: center;
-
-    &-extension {
-      display: flex;
-      margin-left: 1rem;
-    }
-  }
-}
